@@ -1,11 +1,11 @@
 {
-  "name": "angular-maphilight",
-  "version": "0.2.2",
+  "name": "ng-maphilight-demo",
   "license": "ISC",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build-docs": "ng build --base-href /angular-maphilight/ --output-path docs -e prod",
+    "build": "ng build",
+    "build-docs": "ng build --base-href /ng-maphilight/ --output-path docs --prod",
     "build-lib": "bin/build-lib",
     "test": "ng test",
     "lint": "ng lint",
@@ -13,47 +13,48 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/TylerRick/angular-maphilight.git"
+    "url": "git+https://github.com/TylerRick/ng-maphilight.git"
   },
-  "author": "Abdalla Mohamed El-Abd <abdalla.mohamed.elabd@gmail.com>, Tyler Rick",
+  "author": "Abdalla Mohamed El-Abd, Tyler Rick",
   "bugs": {
-    "url": "https://github.com/TylerRick/angular-maphilight/issues"
+    "url": "https://github.com/TylerRick/ng-maphilight/issues"
   },
-  "homepage": "https://github.com/TylerRick/angular-maphilight#readme",
+  "homepage": "https://github.com/TylerRick/ng-maphilight#readme",
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@uirouter/angular": "^1.0.0",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "@angular/animations": "~8.0.1",
+    "@angular/common": "~8.0.1",
+    "@angular/compiler": "~8.0.1",
+    "@angular/core": "~8.0.1",
+    "@angular/forms": "~8.0.1",
+    "@angular/platform-browser": "~8.0.1",
+    "@angular/platform-browser-dynamic": "~8.0.1",
+    "@angular/router": "~8.0.1",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "~8.0.1",
+    "@angular/language-service": "~8.0.1",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ngmakelib": "^0.0.1",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^5.1.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.35.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   }
 }
